+++ tower-http/src/services/fs/serve_dir/tests.rs
@@ -642,6 +642,32 @@ async fn method_not_allowed() {
 }
 
 #[tokio::test]
+async fn calling_fallback_on_not_allowed() {
+    async fn fallback<B>(req: Request<B>) -> io::Result<Response<Body>> {
+        Ok(Response::new(Body::from(format!(
+            "from fallback {}",
+            req.uri().path()
+        ))))
+    }
+
+    let svc = ServeDir::new("..")
+        .call_fallback_on_method_not_allowed(true)
+        .fallback(tower::service_fn(fallback));
+
+    let req = Request::builder()
+        .method(Method::POST)
+        .uri("/doesnt-exist")
+        .body(Body::empty())
+        .unwrap();
+    let res = svc.oneshot(req).await.unwrap();
+
+    assert_eq!(res.status(), StatusCode::OK);
+
+    let body = body_into_text(res.into_body()).await;
+    assert_eq!(body, "from fallback /doesnt-exist");
+}
+
+#[tokio::test]
 async fn with_fallback_svc_and_not_append_index_html_on_directories() {
     async fn fallback<B>(req: Request<B>) -> io::Result<Response<Body>> {
         Ok(Response::new(Body::from(format!(