+++ tower-http/CHANGELOG.md
@@ -9,8 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Added
 
+- **serve_dir:** Add `ServeDir::call_fallback_on_method_not_allowed` to allow calling the fallback
+  for requests that aren't `GET` or `HEAD` ([#264])
 - **request_id:** Add `MakeRequestUuid` for generating request ids using UUIDs ([#266])
 
+[#264]: https://github.com/tower-rs/tower-http/pull/264
 [#266]: https://github.com/tower-rs/tower-http/pull/266
 
 ## Changed